+++ eventdock-core/src/main/java/io/github/oxxultus/eventdock/core/ConsumerMode.java
@@ -1,0 +1,6 @@
+package io.github.oxxultus.eventdock.core;
+
+public enum ConsumerMode {
+  INBOX,
+  DIRECT
+}