+++ eventdock-inbox/src/main/java/io/github/oxxultus/eventdock/inbox/InboxHandlerRegistry.java
@@ -1,6 +1,9 @@
 package io.github.oxxultus.eventdock.inbox;
 
+import io.github.oxxultus.eventdock.core.EventHandlerRegistry;
+
 @FunctionalInterface
-public interface InboxHandlerRegistry {
+public interface InboxHandlerRegistry extends EventHandlerRegistry {
+  @Override
   InboxHandler get(String consumerId, String eventType);
 }